Improvements on settling perp pnl
This is an optimization on f97370f following Dafy's suggestion here:

https://github.com/blockworks-foundation/liquidator-v3/pull/17#issuecomment-1020808416

It uses the client's settlePosPnl function instead of doing all the work
itself. It required making the mangoAccounts array global so we can have
access to it when liquidating from the RPC endpoint, but I'm not getting
any movement from the feed anyway.